Adjusted some stuff
Adjusted the scale of the tray so that it fits better in the hole and changed the material of it to look better. In the future it will need a material created for it. The textures for the inside and outside of the wall need to be edited which Zander said he will do tomorrow.