Fixing references
Somehow all the refrences of objects to each other got reset/ This fixes that. There seems to still be a bug with phone guy out of car animation being broken for some reason which is causing it to look funny and also to happen while the phone is ringing for some reason.